+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # QOwnNotes Changelog
 
 ## 0.61
-- only todo items (not calendar items) are now shown in your todo list
-- you can now choose to hide completed todo items
+- only todo items (not calendar items) are now shown in your todo lists
+- you can now choose to hide completed todo items at the todo list dialog
+- the last selected todo list is now remembered at the todo list dialog
 
 ## 0.60
 - if the `Ctrl` key is pressed in the note text edit to click a link the mouse cursor is now a pointing hand